requirements.txt
telethon
python-dotenv
tqdm
loguru
psycopg2
ultralytics
fastapi
uvicorn
SQLAlchemy